ingester/requirements.txt
appwrite
chromadb
python-dotenv
beautifulsoup4
lxml
tiktoken
requests
ixbrlparse
sec-parser

